// src/app/@core/models/student/student.ts
export interface Student{
	admno?: string;
	admno_order?: number;
	gender?: string;
	isClassMember?: boolean;
	kcpe?: number;
	name?: string;
	phone?: string;
	stream?: string;
	streamid?: number;
	url?: string;
	userid: number;
	present?:number,
	absent?:number
	majorid?: number;
	major?: string;
	birth_cert_number?: string;
	current_class_name?: string;
	date_of_admission: any;
	dob: any;
	form: any;
	general_comments?: string;
	guardian_email?: string;
	guardian_name?: string;
	guardian_relation?: string;
	home_address?: string;
	house?: string;
	indexnumber?: string;
	isKcseSchool?: boolean;
	isClassteacher?: boolean;
	isKcpeSchool?: boolean;
	kcpe_index?: string;
	kcpe_score?: number;
	kcpe_year?:number;
	leftSchool?: boolean;
	leaveDate?: string;
	new_admno?: string;
	new_upi?: string;
	primary_guardian_phone?: string;
	primary_school?: string;
	secondary_guardian_phone?: string;
	upi?: string;
	username?: string;
	enrollmentForm?: string | number;
	nhif?: string;
	boardingStatus?: string;

	upiNumber: string,
	admissionDate?: string;
	graduationDate?: string;
	admissionNumber?: string;
	dateOfBirth?: string;
	graduationForm?: string;
	generalComments?: string;
	issueDate?: string;
}

export class Student {
	admno?: string;
	admno_order?: number;
	gender?: string;
	isClassMember?: boolean;
	kcpe?: number;
	name?: string;
	phone?: string;
	stream?: string;
	streamid?: number;
	url?: string;
	userid = 0;
	present?: number;
	absent?: number;
	majorid?: number;
	birth_cert_number?: string;
	current_class_name?: string;
	date_of_admission: any;
	dob: any;
	form: any;
	general_comments?: string;
	guardian_email?: string;
	guardian_name?: string;
	guardian_relation?: string;
	home_address?: string;
	house?: string;
	indexnumber?: string;
	isKcseSchool?: boolean;
	isClassteacher?: boolean;
	isKcpeSchool?: boolean;
	kcpe_index?: string;
	kcpe_score?: number;
	kcpe_year?:number;
	leftSchool?: boolean;
	leaveDate?: string;
	new_admno?: string;
	new_upi?: string;
	primary_guardian_phone?: string;
	primary_school?: string;
	secondary_guardian_phone?: string;
	upi?: string;
	username?: string;
	enrollmentForm?: string | number;
	nhif?: string;
	boardingStatus?: string;
	upiNumber = "";
	admissionDate?: string;
	graduationDate?: string;
	admissionNumber?: string;
	dateOfBirth?: string;
	graduationForm?: string;
	generalComments?: string;
	issueDate?: string;
}